requirements.txt

streamlit
requests
Pillow
